+++ CHANGES.md
@@ -15,6 +15,8 @@ v1.6.0 - YYYY-MM-DD
 - Added explicit support for warning messages (Issue #118)
 - Added `pdfioFileCreateFontObjFromData` function for embedding fonts in
   memory (Issue #120)
+- Added `pdfioContentBeginMarked` and `pdfioContentEndMarked` functions for
+  creating tagged PDF content (Issue #123)
 - Added `pdfioFileGetLanguage` and `pdfioFileSetLanguage` functions for getting
   and setting the default/primary language of a PDF file (Issue #124)
 - Now add default grayscale, RGB, and CMYK profile resources to pages as needed